// src/app/models/user.model.ts
export class User{
    id: number;
    FirstName: string;
    LastName: string;
    dateOfBirth: Date;
    email: string;
    phoneNumber: number;
    FatLoss: boolean;
    MuscleGain: boolean;
    RehabPrehab: boolean;
    ImproveStrength: boolean;
    SporstSpecific: boolean;
    LearnNewSkill: boolean;
    ImproveFlexibility: boolean;
    Cardiovascular: boolean;
    InfoAboutSpecifical: string;
    BeforeAfter: boolean;
    Weight: boolean;
    BodyFat: boolean;
    TapeMeasure: boolean;
    Performance: boolean;
    Assessments: boolean;
    Achive: string;
    TimeToAchive: string;
    Committed: number;
    Train: string;
    DeatilAboutTrain: string;
    TrainingLike: string;
    Follow: string;
    Coaching: string;
    Problems: string;
    DetailAboutProblems: string;
    Preventing: string;
    Supporting: string;
}